service

+++ utils.py
@@ -498,7 +498,7 @@ class ServiceData:
             for i in df_part.columns:
                 if i in ['소속부문','소속총괄']: df_sum[i] = df_part[i]
                 elif i in ['소속부서']: df_sum[i] = ''
-                else: df_sum[i] = df_part[i].astype(int).sum() # 컬럼별 합계
+                else: df_sum[i] = df_part[i].sum() # 컬럼별 합계
                 # df_sum['소속부문'] = df_part['소속부문']
                 # df_sum['소속총괄'] = df_part['소속총괄']
                 # df_sum['소속부서'] = ''
